+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Irish Beacon Watch Changelog
 ## version 0.5.2
 Added click action to map marker on homepage
+Added Google Custom Search API to Forecast page
 
 ## version 0.5.1
 Added Map to Marine and Historical Pages