+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Irish Beacon Watch Changelog
 
+## version 0.6.1
+Added Country Table and country attribute to Lighthouse object.
+
 ## version 0.6.0
 Moved all weather and marine Forecast into Lightohuse page (Forecast page will now be Lightohuse page)
 Image button unused (swithed off)